AGP wil present and touching screen calls up the timebar, but screen is white and values are incorrect. Also requires a close button
