+++ xdrip/Constants/ConstantsStatistics.swift
@@ -44,11 +44,11 @@ enum ConstantsStatistics {
     static let standardisedHighValueForTITRInMmol = 7.8
     
     // Constants for TIR and TITR  using `UniversalBGLevel`
-    static let standardisedUniLowValueForTITR = UniversalBGLevel(_mgdl: MGDL(ConstantsStatistics.standardisedLowValueForTITRInMgDl))
-    static let standardisedUniHighValueForTITR = UniversalBGLevel(_mgdl: MGDL(ConstantsStatistics.standardisedHighValueForTITRInMgDl))
+    static let standardisedUniLowValueForTITR = UniversalBGLevel(mgdl: MGDL(ConstantsStatistics.standardisedLowValueForTITRInMgDl))
+    static let standardisedUniHighValueForTITR = UniversalBGLevel(mgdl: MGDL(ConstantsStatistics.standardisedHighValueForTITRInMgDl))
     
-    static let standardisedUniLowValueForTIR = UniversalBGLevel(_mgdl: MGDL(ConstantsStatistics.standardisedLowValueForTIRInMgDl))
-    static let standardisedUniHighValueForTIR = UniversalBGLevel(_mgdl: MGDL(ConstantsStatistics.standardisedHighValueForTIRInMgDl))
+    static let standardisedUniLowValueForTIR = UniversalBGLevel(mgdl: MGDL(ConstantsStatistics.standardisedLowValueForTIRInMgDl))
+    static let standardisedUniHighValueForTIR = UniversalBGLevel(mgdl: MGDL(ConstantsStatistics.standardisedHighValueForTIRInMgDl))
     // ----
     
     /// highlight color when changing between TIR modes
